fix(rust, python): is_in struct logical types
