+++ py-polars/tests/unit/datatypes/test_list.py
@@ -247,23 +247,6 @@ def test_list_diagonal_concat() -> None:
     }
 
 
-def test_is_in_empty_list_4559() -> None:
-    assert pl.Series(["a"]).is_in([]).to_list() == [False]
-
-
-def test_is_in_empty_list_4639() -> None:
-    df = pl.DataFrame({"a": [1, None]})
-    empty_list: list[int] = []
-
-    assert df.with_columns([pl.col("a").is_in(empty_list).alias("a_in_list")]).to_dict(
-        False
-    ) == {"a": [1, None], "a_in_list": [False, False]}
-    df = pl.DataFrame()
-    assert df.with_columns(
-        [pl.lit(None).cast(pl.Int64).is_in(empty_list).alias("in_empty_list")]
-    ).to_dict(False) == {"in_empty_list": [False]}
-
-
 def test_inner_type_categorical_on_rechunk() -> None:
     df = pl.DataFrame({"cats": ["foo", "bar"]}).select(
         pl.col(pl.Utf8).cast(pl.Categorical).implode()
