+++ py-polars/tests/unit/datatypes/test_struct.py
@@ -557,23 +557,6 @@ def test_arr_unique() -> None:
     assert {"a": 1, "b": 11} in unique_el
 
 
-def test_is_in_struct() -> None:
-    df = pl.DataFrame(
-        {
-            "struct_elem": [{"a": 1, "b": 11}, {"a": 1, "b": 90}],
-            "struct_list": [
-                [{"a": 1, "b": 11}, {"a": 2, "b": 12}, {"a": 3, "b": 13}],
-                [{"a": 3, "b": 3}],
-            ],
-        }
-    )
-
-    assert df.filter(pl.col("struct_elem").is_in("struct_list")).to_dict(False) == {
-        "struct_elem": [{"a": 1, "b": 11}],
-        "struct_list": [[{"a": 1, "b": 11}, {"a": 2, "b": 12}, {"a": 3, "b": 13}]],
-    }
-
-
 def test_nested_explode_4026() -> None:
     df = pl.DataFrame(
         {
